Only provision enabled settings

+++ config/plex/provision.sh
@@ -1,8 +1,23 @@
 #!/bin/bash
 
-echo -e "\nProvisioning plex"
 source ./.env
 
-tar -xzf ./config/plex/provision.tar.gz -C $DIR_DATA_ROOT
-cp ./config/plex/Preferences.xml.template "$DIR_DATA_ROOT/plex/Library/Application Support/Plex Media Server/Preferences.xml"
-sed -i "s/\${EXTERNAL_DOMAIN}/${EXTERNAL_DOMAIN}/g" "$DIR_DATA_ROOT/plex/Library/Application Support/Plex Media Server/Preferences.xml"
+echo -e "\nPlex Setup"
+read -p "Visit https://plex.tv/claim and paste claim token, or press enter to skip: " PLEX_CLAIM
+
+if [ -z "$PLEX_CLAIM" ]; then
+    #Disable plex in .env
+    echo "Skipping Plex setup"
+    ENABLE_PLEX=0
+    sed -i "s|^ENABLE_PLEX=.*|ENABLE_PLEX=\"$ENABLE_PLEX\"|" .env
+else
+    #Set the claim token
+    sed -i "s|^PLEX_CLAIM=.*|PLEX_CLAIM=$PLEX_CLAIM|" .env
+    #Provision default settings
+    tar -xzf ./config/plex/provision.tar.gz -C $DIR_DATA_ROOT
+    cp ./config/plex/Preferences.xml.template "$DIR_DATA_ROOT/plex/Library/Application Support/Plex Media Server/Preferences.xml"
+    sed -i "s/\${EXTERNAL_DOMAIN}/${EXTERNAL_DOMAIN}/g" "$DIR_DATA_ROOT/plex/Library/Application Support/Plex Media Server/Preferences.xml"
+fi
+
+echo -e "\Plex Setup complete"
+
